feat: add behaviour to delete and re-create a vesting schedule
This happens when a schedule needs to be updated but the original schedule is claimable, unclaimed and end date passed. Those schedules cant be updated anymore, so the behavior is to delete those schedules and create new ones with the updated total amount. The amount from previous schedule is transferred as cliff.